Move remaining exceptions to physics
