+++ plasmapy/physics/exceptions.py
