+++ plasmapy/physics/tests/test_quantum.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 import astropy.units as u
 from ...constants import c, h
-from ...utils.exceptions import RelativityError
+from plasmapy.physics.exceptions import RelativityError
 from ..quantum import (deBroglie_wavelength,
                        thermal_deBroglie_wavelength,
                        Fermi_energy,
