+++ plasmapy/physics/tests/test_relativity.py
@@ -5,7 +5,7 @@ import numpy as np
 from astropy import units as u
 from ...constants import c
 from ..relativity import Lorentz_factor
-from ...utils.exceptions import RelativityError
+from plasmapy.physics.exceptions import RelativityError
 
 
 def test_Lorentz_factor():
