+++ plasmapy/transport/tests/test_collisions.py
@@ -15,9 +15,9 @@ from plasmapy.transport import (Coulomb_logarithm,
 from plasmapy.transport import (Spitzer_resistivity,
                                 fundamental_electron_collision_freq,
                                 fundamental_ion_collision_freq)
-from plasmapy.utils import exceptions
+from plasmapy.physics import exceptions
 from plasmapy.utils.pytest_helpers import assert_can_handle_nparray
-from plasmapy.constants import m_p, m_e, c
+from plasmapy.constants import c
 
 
 class Test_Coulomb_logarithm:
