+++ plasmapy/transport/tests/test_transport.py
@@ -5,9 +5,9 @@ import pytest
 from astropy import units as u
 from plasmapy.atomic.atomic import particle_mass, integer_charge
 from astropy.tests.helper import assert_quantity_allclose
-from plasmapy.utils.exceptions import (PhysicsError,
-                                       PhysicsWarning,
-                                       RelativityWarning)
+from plasmapy.physics.exceptions import (PhysicsError,
+                                         PhysicsWarning,
+                                         RelativityWarning)
 from plasmapy.atomic._exceptions import InvalidParticleError
 from plasmapy.physics.parameters import Hall_parameter
 from plasmapy.constants import m_p, m_e
